# docs/swagger.yaml
basePath: /
definitions:
  github_com_cdcdx_high-concurrency-framework_internal_model.ApiResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      trace_id:
        type: string
    type: object
  github_com_cdcdx_high-concurrency-framework_internal_model.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
  github_com_cdcdx_high-concurrency-framework_internal_model.RegisterRequest:
    properties:
      email:
        maxLength: 128
        type: string
      password:
        maxLength: 128
        minLength: 6
        type: string
      phone:
        maxLength: 20
        type: string
      username:
        maxLength: 64
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
  github_com_cdcdx_high-concurrency-framework_internal_model.TokenResponse:
    properties:
      access_token:
        type: string
      expires_in:
        description: 秒
        type: integer
      token_type:
        description: '"Bearer"'
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
  github_com_cdcdx_high-concurrency-framework_internal_model.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      last_login_at:
        type: string
      phone:
        type: string
      status:
        description: 1=正常 0=禁用
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: cdcdx
  description: 高并发业务处理框架，支持多级缓存、读写分离、异步消息、全文搜索等能力
  title: 高并发业务框架 API
  version: 1.0.0
paths:
  /api/v1/analytics/behaviors:
    get:
      description: 查询用户行为事件汇总数据
      parameters:
      - description: 事件类型(如order_create)
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_cdcdx_high-concurrency-framework_internal_model.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_cdcdx_high-concurrency-framework_internal_model.ApiResponse'
      security:
      - Bearer: []
      summary: 行为摘要
      tags:
      - 分析
  /api/v1/analytics/daily:
    get:
      description: 查询PostgreSQL中的每日统计数据，支持日期范围筛选
      parameters:
      - default: __7DAYS_AGO__
        description: 起始日期(YYYY-MM-DD)
        in: query
        name: from
        type: string
      - default: __TODAY__
        description: 结束日期(YYYY-MM-DD)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_cdcdx_high-concurrency-framework_internal_model.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_cdcdx_high-concurrency-framework_internal_model.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_cdcdx_high-concurrency-framework_internal_model.ApiResponse'
      security:
      - Bearer: []
      summary: 每日统计
      tags:
      - 分析
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: 验证用户名密码并返回JWT令牌
      parameters:
      - description: 登录信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_cdcdx_high-concurrency-framework_internal_model.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_cdcdx_high-concurrency-framework_internal_model.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_cdcdx_high-concurrency-framework_internal_model.TokenResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_cdcdx_high-concurrency-framework_internal_model.ApiResponse'
      summary: 用户登录
      tags:
      - 认证
  /api/v1/auth/me:
    get:
      description: 根据JWT令牌返回当前登录用户信息 (需Authorization头)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_cdcdx_high-concurrency-framework_internal_model.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_cdcdx_high-concurrency-framework_internal_model.User'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_cdcdx_high-concurrency-framework_internal_model.ApiResponse'
      security:
      - Bearer: []
      summary: 获取当前用户信息
      tags:
      - 认证
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: 创建新用户账号并返回JWT令牌
      parameters:
      - description: 注册信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_cdcdx_high-concurrency-framework_internal_model.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_cdcdx_high-concurrency-framework_internal_model.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_cdcdx_high-concurrency-framework_internal_model.TokenResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_cdcdx_high-concurrency-framework_internal_model.ApiResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_cdcdx_high-concurrency-framework_internal_model.ApiResponse'
      summary: 用户注册
      tags:
      - 认证
...
securityDefinitions:
  Bearer:
    description: 输入JWT令牌 (可带或不带 "Bearer " 前缀), 通过 POST /api/v1/auth/login 或 /api/v1/auth/register
      获取
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"